requirements.txt
anyio==3.6.2
certifi==2023.5.7
distlib==0.3.6
filelock==3.12.0
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
idna==3.4
platformdirs==3.5.1
python-telegram-bot==20.3
sniffio==1.3.0
virtualenv==20.23.0
